Create .gitignore
The purpose of this gitignore is to ignore everything except specific whitelisted files.  This means the repo can be synced with a home folder and ONLY the files named here will sync (lest we upload all our personal data).  To add a file simply name it.  To add a folder name it and end it with '/*' to include all its contents.